QMUI升级到2.0

+++ QMUI/QMUIKit/UIComponents/QMUIPieProgressView.m
@@ -7,8 +7,7 @@
 //
 
 #import "QMUIPieProgressView.h"
-#import "QMUICommonDefines.h"
-#import "QMUIConfigurationMacros.h"
+#import "QMUICore.h"
 
 @interface QMUIPieProgressLayer : CALayer
 
@@ -109,7 +108,7 @@
 }
 
 - (void)setProgress:(float)progress animated:(BOOL)animated {
-    _progress = fmaxf(0.0, fminf(1.0, progress));
+    _progress = fmax(0.0, fmin(1.0, progress));
     QMUIPieProgressLayer *layer = (QMUIPieProgressLayer *)self.layer;
     layer.shouldChangeProgressWithAnimation = animated;
     layer.progress = _progress;
